move /cache & /api-docs to serviceRoutes
