+++ src/routes/serviceRoutes.js
@@ -1,6 +1,11 @@
 const router = require('express').Router();
+const swaggerUi = require('swagger-ui-express');
+const apicache = require('apicache');
+
+const { client } = require('../services/RedisService');
 const ResponseHandler = require('../utils/ResponseHandler');
-const Constants = require('../constants/Constants');
+const apiSpecification = require('./utils/swaggerConfig');
+const { NAME, VERSION, ENVIRONMENT } = require('../constants/Constants');
 
 /**
  * @openapi
@@ -13,13 +18,53 @@ const Constants = require('../constants/Constants');
  *         description: Returns a mysterious JSON.
  */
 router.get('/', (_, res) => {
-  const { NAME, ENVIRONMENT, VERSION } = Constants;
-
   ResponseHandler.sendRes(res, {
     environment: ENVIRONMENT,
     name: NAME,
     version: VERSION,
   });
+});
+
+/**
+ * @openapi
+ * /api-docs:
+ *   get:
+ *     description: Documentación de rutas de la api
+ *     responses:
+ *       200:
+ *         description: Returns a mysterious webpage.
+ */
+router.use('/api-docs', swaggerUi.serve, swaggerUi.setup(apiSpecification));
+
+router.use(
+  apicache
+    .options({
+      redisClient: client,
+      debug: false,
+      trackPerformance: true,
+      respectCacheControl: false,
+    })
+    .middleware(),
+);
+
+/**
+ * @openapi
+ * /cache:
+ *   get:
+ *     description: Información sobre la cache
+ *     responses:
+ *       200:
+ *         description: Returns a mysterious webpage.
+ */
+router.get('/cache', (_, res) => {
+  try {
+    res.json({
+      performance: apicache.getPerformance(),
+      index: apicache.getIndex(),
+    });
+  } catch (e) {
+    res.status(500);
+  }
 });
 
 module.exports = router;
